server side xray

+++ src/AppGet/FileTransfer/Protocols/HttpFileTransferClient.cs
@@ -15,15 +15,17 @@ namespace AppGet.FileTransfer.Protocols
 {
     public class HttpFileTransferClient : IFileTransferClient
     {
-        private readonly IHttpClient _httpClient;
+        private static readonly Regex HttpRegex =
+            new Regex(@"^https?\:\/\/", RegexOptions.Compiled | RegexOptions.IgnoreCase);
+
+        private static readonly Regex FileNameRegex = new Regex(@"\.(zip|7zip|7z|rar|msi|exe)$",
+            RegexOptions.Compiled | RegexOptions.IgnoreCase);
+
+        private static readonly Dictionary<string, WebHeaderCollection> HeaderCache =
+            new Dictionary<string, WebHeaderCollection>();
+
         private readonly IFileSystem _fileSystem;
-        private static readonly Regex HttpRegex = new Regex(@"^https?\:\/\/", RegexOptions.Compiled | RegexOptions.IgnoreCase);
-        private static readonly Regex FileNameRegex = new Regex(@"\.(zip|7zip|7z|rar|msi|exe)$", RegexOptions.Compiled | RegexOptions.IgnoreCase);
-
-
-
-        private static readonly Dictionary<string, WebHeaderCollection> HeaderCache = new Dictionary<string, WebHeaderCollection>();
-
+        private readonly IHttpClient _httpClient;
 
         public HttpFileTransferClient(IHttpClient httpClient, IFileSystem fileSystem)
         {
@@ -36,32 +38,23 @@ namespace AppGet.FileTransfer.Protocols
             return HttpRegex.IsMatch(source);
         }
 
-
         public async Task<string> GetFileName(string source)
         {
             var uri = new Uri(source);
 
             var fileName = Path.GetFileName(uri.LocalPath);
 
-            if (FileNameRegex.IsMatch(fileName))
-            {
-                return fileName;
-            }
+            if (FileNameRegex.IsMatch(fileName)) return fileName;
 
-            var resp = await _httpClient.Head(uri);
+            var resp = await _httpClient.Get(uri, HttpCompletionOption.ResponseHeadersRead);
 
             if (resp.RequestMessage.RequestUri != uri)
-            {
                 return await GetFileName(resp.RequestMessage.RequestUri.ToString());
-            }
 
             if (resp.Content.Headers.ContentDisposition != null)
-            {
                 return resp.Content.Headers.ContentDisposition.FileName.Trim('"', '\'', ' ');
-            }
 
             throw new InvalidDownloadUrlException(source);
-
         }
 
         public void TransferFile(string source, string destinationFile)
@@ -72,32 +65,21 @@ namespace AppGet.FileTransfer.Protocols
 
             using (var webClient = new WebClientWithTimeout(TimeSpan.FromSeconds(10)))
             {
-
-                if (_fileSystem.FileExists(destinationFile))
-                {
-                    _fileSystem.DeleteFile(destinationFile);
-                }
+                if (_fileSystem.FileExists(destinationFile)) _fileSystem.DeleteFile(destinationFile);
 
                 webClient.DownloadProgressChanged += (sender, e) =>
                 {
                     progress.Completed = e.BytesReceived;
 
                     if (e.TotalBytesToReceive > 0)
-                    {
                         progress.Total = e.TotalBytesToReceive;
-                    }
                     else
-                    {
                         progress.Total = null;
-                    }
                 };
 
                 webClient.DownloadFileCompleted += (sender, e) =>
                 {
-                    if (error == null)
-                    {
-                        error = e.Error;
-                    }
+                    if (error == null) error = e.Error;
 
                     if (error != null) return;
 
@@ -105,9 +87,7 @@ namespace AppGet.FileTransfer.Protocols
 
                     var contentType = client.ResponseHeaders["Content-Type"];
                     if (contentType != null && contentType.Contains("text"))
-                    {
                         error = new InvalidDownloadUrlException(client.BaseAddress, $"[ContentType={contentType}]");
-                    }
                 };
 
                 webClient.DownloadFileAsync(new Uri(source), tempFile);
@@ -120,10 +100,7 @@ namespace AppGet.FileTransfer.Protocols
 
                 if (error != null)
                 {
-                    if (_fileSystem.FileExists(tempFile))
-                    {
-                        _fileSystem.DeleteFile(tempFile);
-                    }
+                    if (_fileSystem.FileExists(tempFile)) _fileSystem.DeleteFile(tempFile);
 
                     throw error;
                 }
@@ -144,13 +121,13 @@ namespace AppGet.FileTransfer.Protocols
             return await resp.Content.ReadAsStringAsync();
         }
 
+        public Action<ProgressState> OnStatusUpdated { get; set; }
+        public Action<ProgressState> OnCompleted { get; set; }
+
 
         public static WebHeaderCollection GetTransferHeaders(string url)
         {
             return HeaderCache.ContainsKey(url) ? HeaderCache[url] : null;
         }
-
-        public Action<ProgressState> OnStatusUpdated { get; set; }
-        public Action<ProgressState> OnCompleted { get; set; }
     }
 }
